move util function to a file
